@@ -10,20 +10,25 @@
         "production": "mix --production"
     },
     "devDependencies": {
-        "@popperjs/core": "^2.11.2",
-        "axios": "^0.25",
+        "@inertiajs/inertia": "^0.10.0",
+        "@inertiajs/inertia-vue3": "^0.5.1",
+        "@inertiajs/progress": "^0.2.6",
+        "@popperjs/core": "^2.10.2",
+        "@tailwindcss/forms": "^0.2.1",
+        "@tailwindcss/typography": "^0.3.0",
+        "@vue/compiler-sfc": "^3.0.5",
+        "axios": "^0.21.4",
         "bootstrap": "^5.1.3",
-        "laravel-mix": "^6.0.41",
-        "lodash": "^4.17.21",
-        "postcss": "^8.4.5",
+        "css-loader": "^6.5.1",
+        "laravel-mix": "^6.0.6",
+        "lodash": "^4.17.19",
+        "postcss": "^8.1.14",
+        "postcss-import": "^14.0.2",
         "resolve-url-loader": "^5.0.0",
-        "sass": "^1.49.0",
-        "sass-loader": "^12.4.0",
-        "vue": "^2.6.14",
-        "vue-loader": "^17.0.0",
-        "vue-template-compiler": "^2.6.14"
-    },
-    "dependencies": {
-        "vue-router": "^3.5.3"
+        "sass": "^1.32.11",
+        "sass-loader": "^11.1.1",
+        "tailwindcss": "^2.0.1",
+        "vue": "^3.0.5",
+        "vue-loader": "^16.8.3"
     }
 }
